requirements.txt
kafka-python==2.0.2
numpy==1.22.2
opencv-python==4.5.5.62
